Include hadronization hadrons in afterburner
This includes the fragmentation hadrons from hadronization modules into 
the afterburner calculation. This only works if the hadron have a 
position and therefore will only run with Hybrid Hadronization. But 
would run with other Hadronization modules if they put position for 
hadrons in the future.

Also included is a much more useful Afterburner class that takes care of 
gathering all hadrons (soft or hard) to be put into any Afterburner run.
Repeated sampling for soft hadrons is blocked in case of including 
fragmentation hadrons, to match hadronization events with soft 
particlization events.